+++ completion/git-machete.completion.bash
@@ -17,10 +17,10 @@ _git_machete() {
     discover_opts="-C --checked-out-since= -l --list-commits -r --roots= -y --yes"
     fork_point_opts="--inferred --override-to= --override-to-inferred --override-to-parent --unset-override"
     reapply_opts="-f --fork-point="
-    slide_out_opts="-d --down-fork-point= -n --no-interactive-rebase"
+    slide_out_opts="-d --down-fork-point= -M --merge -n --no-edit-merge --no-interactive-rebase"
     status_opts="--color= -L --list-commits-with-hashes -l --list-commits"
-    traverse_opts="-F --fetch -l --list-commits -n --no-interactive-rebase --return-to= --start-from= -w --whole -W --yes"
-    update_opts="-f --fork-point= -n --no-interactive-rebase"
+    traverse_opts="-F --fetch -l --list-commits -M --merge -n --no-edit-merge --no-interactive-rebase --return-to= --start-from= -w --whole -W --yes"
+    update_opts="-f --fork-point= -M --merge -n --no-edit-merge --no-interactive-rebase"
 
     case $cur in
         --checked-out-since=*) __gitcomp "" ;;